+++ packages/network/lib/src/exception.dart
@@ -1,8 +1,12 @@
 import 'package:network/src/response.dart';
 
 class NetworkException<T extends BinaryResponse> implements Exception {
-  final T response;
   NetworkException(this.response);
 
+  final T response;
+
   int get code => response.statusCode;
+
+  @override
+  String toString() => '$runtimeType{code: $code}';
 }