+++ packages/network/lib/src/methods/put.dart
@@ -4,7 +4,7 @@ import 'dart:convert' show Encoding, jsonEncode;
 import 'package:http/http.dart' as http show put, Response;
 import 'package:network/src/exception.dart';
 import 'package:network/src/response.dart';
-import 'package:network/src/utils.dart';
+import 'package:network/src/utils/response_by_type.dart';
 import 'package:network/src/utils/serialize_query_params.dart';
 
 Future<T> put<T extends BinaryResponse>(